order service update

+++ src/main/java/com/example/ecommerce/service/PaymentService.java
@@ -1,5 +1,7 @@
 package com.example.ecommerce.service;
 
+import java.math.BigDecimal;
+
 public interface PaymentService {
-    public boolean processPayment(String orderID, String paymentMethod, double amount);
+    public boolean processPayment(String orderID, String paymentMethod, BigDecimal amount);
 }
