+++ src/main/java/com/example/ecommerce/service/impl/PaymentServiceImpl.java
@@ -6,12 +6,14 @@ import com.example.ecommerce.service.PaymentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 @Service
 @RequiredArgsConstructor
 public class PaymentServiceImpl implements PaymentService {
     private final PaymentRepository paymentRepository;
 
-    public boolean processPayment(String orderID, String paymentMethod, double amount){
+    public boolean processPayment(String orderID, String paymentMethod, BigDecimal amount){
         PaymentMethod paymentMethodEnum = PaymentMethod.valueOf(paymentMethod);
         //save payment details in DB
         return true;
